# documentation/api/identity.openapi.yaml
openapi: 3.1.0
info:
  title: Zagrosi Identity API
  version: 0.1.0
  summary: Public identity, SSO, SCIM, and token-management API.
  description: >
    This specification documents the browser auth, OIDC, SAML, SCIM,
    session, personal access token, SCIM token, service token, and admin
    unlock surfaces shipped by the identity foundation.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
...
servers:
  - url: https://api.zagrosi.example
    description: Production API gateway
security: []
...
paths:
...
  /scim/v2/Groups:
    get:
      tags: [SCIM]
      operationId: scimListGroups
      summary: List SCIM groups.
      security:
        - ScimBearer: []
      responses:
        "200":
          $ref: "#/components/responses/ScimList"
    post:
      tags: [SCIM]
      operationId: scimCreateGroup
      summary: Create a SCIM group.
      security:
        - ScimBearer: []
      requestBody:
        $ref: "#/components/requestBodies/ScimGroup"
      responses:
        "201":
          $ref: "#/components/responses/ScimResource"
  /scim/v2/Groups/{id}:
    get:
      tags: [SCIM]
      operationId: scimGetGroup
      summary: Fetch a SCIM group.
      security:
        - ScimBearer: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "200":
          $ref: "#/components/responses/ScimResource"
    put:
      tags: [SCIM]
      operationId: scimPutGroup
      summary: Replace a SCIM group.
      security:
        - ScimBearer: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        $ref: "#/components/requestBodies/ScimGroup"
      responses:
        "200":
          $ref: "#/components/responses/ScimResource"
    patch:
      tags: [SCIM]
      operationId: scimPatchGroup
      summary: Patch a SCIM group.
      security:
        - ScimBearer: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        $ref: "#/components/requestBodies/ScimPatch"
      responses:
        "200":
          $ref: "#/components/responses/ScimResource"
    delete:
      tags: [SCIM]
      operationId: scimDeleteGroup
      summary: Delete a SCIM group.
      security:
        - ScimBearer: []
      parameters:
        - $ref: "#/components/parameters/UuidId"
      responses:
        "204":
          description: Group deleted.
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: __Host-zagrosi_sid
      description: Secure HttpOnly SameSite=Lax Path=/ session cookie.
    CsrfCookie:
      type: apiKey
      in: cookie
      name: __Host-zagrosi_csrf
      description: Double-submit CSRF cookie paired with `X-Zagrosi-CSRF`.
    OidcCsrfCookie:
      type: apiKey
      in: cookie
      name: __Host-zagrosi_oidc_csrf
      description: OIDC callback state cookie.
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: "pat_<43>, scim_<43>, svc_<43>"
    ScimBearer:
      type: http
      scheme: bearer
      bearerFormat: scim_<43>
  parameters:
    UuidId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IfMatch:
      name: If-Match
      in: header
      required: false
      schema:
        type: string
      description: SCIM ETag precondition per RFC 7644 section 3.14.
...
  headers:
...
    ETag:
      schema:
        type: string
      description: SCIM `meta.version` ETag.
  requestBodies:
...
    ScimGroup:
      required: true
      content:
        application/scim+json:
          schema:
            $ref: "#/components/schemas/ScimGroup"
    ScimPatch:
      required: true
      content:
        application/scim+json:
          schema:
            $ref: "#/components/schemas/ScimPatchRequest"
  responses:
...
    ScimList:
      description: SCIM ListResponse.
      headers:
        ETag:
          $ref: "#/components/headers/ETag"
      content:
        application/scim+json:
          schema:
            $ref: "#/components/schemas/ScimListResponse"
    ScimResource:
      description: SCIM resource.
      headers:
        ETag:
          $ref: "#/components/headers/ETag"
      content:
        application/scim+json:
          schema:
            type: object
            additionalProperties: true
...
  schemas:
...
    ScimGroup:
      type: object
      additionalProperties: true
    ScimPatchRequest:
      type: object
      required: [Operations]
      properties:
        Operations:
          type: array
          items:
            type: object
            additionalProperties: true
    ScimListResponse:
      type: object
      required: [schemas, totalResults, Resources]
      properties:
        schemas:
          type: array
          items:
            type: string
        totalResults:
          type: integer
        startIndex:
          type: integer
        itemsPerPage:
          type: integer
        Resources:
          type: array
          items:
            type: object
            additionalProperties: true
...